requirements.txt
# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2

# Vector Store
chromadb==0.4.22

# PDF Processing
pypdf==4.0.1
pdfplumber==0.10.3

# OpenAI
openai==2.0.1

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.2

# Optional but helpful
tqdm==4.66.1
colorama==0.4.6